Updates omnichannel-server to api Creates server directory to re-organize server logic
