+++ api/src/feature/message-source/util/MessageSourceDigest.ts
