+++ api/src/feature/onboarding/controller/OnboardingController.ts
@@ -1,0 +1,5 @@
+export class OnboardingController {
+  async sendPhoneNumberVerificationCode(input): Promise<boolean> {
+    return false;
+  }
+}
