+++ api/src/feature/onboarding/resolver/sendPhoneNumberVerificationCode.ts
@@ -1,17 +1,16 @@
+import { OnboardingControllerInjectionKey } from "../InjectionKeys";
+
 export const sendPhoneNumberVerificationCode = (
   parent,
   input,
   { dependencies },
   info
-) => {
+): Promise<boolean> => {
   try {
     const onboardingController = dependencies.provide(
-      OnboardingRepositoryInjectionKey
+      OnboardingControllerInjectionKey
     );
 
     return onboardingController.sendPhoneNumberVerificationCode(input);
-    // const response = { status: "success" };
-
-    // return response.status === "success";
   } catch (error) {}
 };
