+++ api/src/index.ts
@@ -1,0 +1,5 @@
+import app from "./server/apollo";
+
+app.listen("8001", () => {
+  console.log("listening");
+});
