+++ api/src/server/resolvers.ts
@@ -1,0 +1,7 @@
+import { queries as onboardingQueries } from "../feature/onboarding/resolver";
+
+export const resolvers = {
+  Query: {
+    ...onboardingQueries,
+  },
+};
