--- omnichannel-server/src/index.ts
@@ -1,5 +1,0 @@
-import server from "./server";
-
-server.listen("8001", () => {
-  console.log("listening");
-});
